feat: add ability to send to ftps server
This commit adds a new shell script called `sendtoftps.sh` that uses
curl to send a file to an ftps server.  I've also added it to all of the
scanto* shell scripts.  If you don't pass in the FTP env vars it will
simply skip the step where it uploads via curl.